# internal/swagger/swagger.yaml
basePath: /
definitions:
...
  controller.OpenIDConnectConfiguration:
    properties:
      authorization_endpoint:
        type: string
      claims_supported:
        items:
          type: string
        type: array
      grant_types_supported:
        items:
          type: string
        type: array
      id_token_signing_alg_values_supported:
        items:
          type: string
        type: array
      issuer:
        type: string
      jwks_uri:
        type: string
      request_object_signing_alg_values_supported:
        items:
          type: string
        type: array
      request_parameter_supported:
        type: boolean
      response_types_supported:
        items:
          type: string
        type: array
      scopes_supported:
        items:
          type: string
        type: array
      service_documentation:
        type: string
      subject_types_supported:
        items:
          type: string
        type: array
      token_endpoint:
        type: string
      token_endpoint_auth_methods_supported:
        items:
          type: string
        type: array
      userinfo_endpoint:
        type: string
    type: object
  controller.SimpleResponse:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
  controller.WebfingerResponse:
    properties:
      links:
        items:
          $ref: '#/definitions/controller.WebfingerResponseLink'
        type: array
      subject:
        type: string
    type: object
  controller.WebfingerResponseLink:
    properties:
      href:
        type: string
      rel:
        type: string
    type: object
...
info:
  contact: {}
  description: Swagger documentation for Tinyauth's API.
  license:
    name: AGPL-3.0
    url: https://github.com/tinyauthapp/tinyauth/blob/main/LICENSE
  title: Tinyauth API
  version: development
paths:
  /.well-known/jwks.json:
    get:
      description: JWKS Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: JWKS
      tags:
      - well-known
  /.well-known/openid-configuration:
    get:
      description: OpenID Connect Configuration Discovery Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.OpenIDConnectConfiguration'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: OpenID Connect Configuration
      tags:
      - well-known
  /.well-known/webfinger:
    get:
      description: WebFinger Endpoint
      parameters:
      - description: Resource
        in: query
        name: resource
        required: true
        type: string
      - description: Rel
        in: query
        name: rel
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.WebfingerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: WebFinger
      tags:
      - well-known
...
swagger: "2.0"